int conversion for attention inner ratio

+++ multi_sources/models/perceiver/cross_att.py
@@ -54,7 +54,7 @@ class MultisourcePerceiverEncoder(nn.Module):
         super().__init__()
         self.values_dim = values_dim
         self.coords_dim = coords_dim
-        self.att_inner_dim = att_inner_ratio * values_dim
+        self.att_inner_dim = int(att_inner_ratio * values_dim)
         self.head_dim = self.att_inner_dim // num_heads
         self.num_heads = num_heads
 
@@ -193,7 +193,7 @@ class MultisourcePerceiverDecoder(nn.Module):
         super().__init__()
         self.values_dim = values_dim
         self.coords_dim = coords_dim
-        self.att_inner_dim = att_inner_ratio * values_dim
+        self.att_inner_dim = int(att_inner_ratio * values_dim)
         self.head_dim = self.att_inner_dim // num_heads
         self.num_heads = num_heads
